+++ app/serializers/character_index_serializer.rb
@@ -1,3 +1,7 @@
 class CharacterIndexSerializer < ActiveModel::Serializer
-  attributes :id, :name
+  attributes :id, :name, :fraction_name
+
+  def fraction_name
+    object.race.fraction.name['en']
+  end
 end