+++ app/serializers/event_edit_serializer.rb
@@ -1,5 +1,5 @@
 class EventEditSerializer < ActiveModel::Serializer
-  attributes :id, :name, :date, :time, :slug, :fraction_id, :description, :dungeon_id, :owner_id, :event_type, :eventable_type, :eventable_id, :group_role
+  attributes :id, :name, :date, :time, :slug, :fraction_id, :description, :dungeon_id, :owner_id, :event_type, :eventable_type, :eventable_id, :group_role, :fraction_name
 
   def date
     object.start_time.strftime('%-d.%-m.%Y')
@@ -10,6 +10,10 @@ class EventEditSerializer < ActiveModel::Serializer
       hours: object.start_time.strftime('%H').to_i,
       minutes: object.start_time.strftime('%M').to_i
     }
+  end
+
+  def fraction_name
+    object.fraction.name
   end
 
   def group_role