+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Modified
+- hide paladin/shaman for static/events create/edit forms
+
 ### Fixed
 - redirect after upload recipes for characters
 - make comment in static
 - craft bug based on old cache value for guilds
 - deleting guild when no members
+- bug with creating character without guild invite
 
 ## [0.9.9] - 2019-12-07
 ### Added